Endrer soft delete mutations til å gjøre hard delete events. Vi har besluttet og varslet til alle produsenter at soft delete fases ut. All sletting vil være hard delete. Første steg er å endre mutations slik at kun hard delete events blir produsert. Senere vil soft delete mutations fjernes.
