+++ app/src/main/kotlin/no/nav/arbeidsgiver/notifikasjon/produsent/api/Error.kt
@@ -12,10 +12,10 @@ internal sealed class Error {
         MutationNyBeskjed.NyBeskjedResultat,
         MutationNyOppgave.NyOppgaveResultat,
         MutationNySak.NySakResultat,
-        MutationHardDeleteNotifikasjon.HardDeleteNotifikasjonResultat,
-        MutationSoftDeleteNotifikasjon.SoftDeleteNotifikasjonResultat,
-        MutationHardDeleteSak.HardDeleteSakResultat,
-        MutationSoftDeleteSak.SoftDeleteSakResultat,
+        MutationDelete.HardDeleteNotifikasjonResultat,
+        MutationDelete.SoftDeleteNotifikasjonResultat,
+        MutationDelete.HardDeleteSakResultat,
+        MutationDelete.SoftDeleteSakResultat,
         MutationKalenderavtale.NyKalenderavtaleResultat
 
     @JsonTypeName("UgyldigMerkelapp")
@@ -30,10 +30,10 @@ internal sealed class Error {
         QueryNotifikasjoner.MineNotifikasjonerResultat,
         QueryNotifikasjoner.HentNotifikasjonResultat,
         QuerySak.HentSakResultat,
-        MutationSoftDeleteSak.SoftDeleteSakResultat,
-        MutationHardDeleteSak.HardDeleteSakResultat,
-        MutationSoftDeleteNotifikasjon.SoftDeleteNotifikasjonResultat,
-        MutationHardDeleteNotifikasjon.HardDeleteNotifikasjonResultat,
+        MutationDelete.SoftDeleteSakResultat,
+        MutationDelete.HardDeleteSakResultat,
+        MutationDelete.SoftDeleteNotifikasjonResultat,
+        MutationDelete.HardDeleteNotifikasjonResultat,
         MutationNyStatusSak.NyStatusSakResultat,
         MutationNesteStegSak.NesteStegSakResultat,
         MutationKalenderavtale.OppdaterKalenderavtaleResultat,
@@ -52,10 +52,10 @@ internal sealed class Error {
         QueryNotifikasjoner.MineNotifikasjonerResultat,
         QueryNotifikasjoner.HentNotifikasjonResultat,
         QuerySak.HentSakResultat,
-        MutationSoftDeleteSak.SoftDeleteSakResultat,
-        MutationHardDeleteSak.HardDeleteSakResultat,
-        MutationSoftDeleteNotifikasjon.SoftDeleteNotifikasjonResultat,
-        MutationHardDeleteNotifikasjon.HardDeleteNotifikasjonResultat,
+        MutationDelete.SoftDeleteSakResultat,
+        MutationDelete.HardDeleteSakResultat,
+        MutationDelete.SoftDeleteNotifikasjonResultat,
+        MutationDelete.HardDeleteNotifikasjonResultat,
         MutationNySak.NySakResultat,
         MutationNyStatusSak.NyStatusSakResultat,
         MutationNesteStegSak.NesteStegSakResultat,
@@ -111,8 +111,8 @@ internal sealed class Error {
         MutationOppgaveUtfoert.OppgaveUtfoertResultat,
         MutationOppgaveUtgaatt.OppgaveUtgaattResultat,
         MutationOppgaveUtsettFrist.OppgaveUtsettFristResultat,
-        MutationSoftDeleteNotifikasjon.SoftDeleteNotifikasjonResultat,
-        MutationHardDeleteNotifikasjon.HardDeleteNotifikasjonResultat,
+        MutationDelete.SoftDeleteNotifikasjonResultat,
+        MutationDelete.HardDeleteNotifikasjonResultat,
         QueryNotifikasjoner.HentNotifikasjonResultat,
         MutationKalenderavtale.OppdaterKalenderavtaleResultat,
         MutationOppgavePåminnelse.OppgaveEndrePaaminnelseResultat
@@ -139,8 +139,8 @@ internal sealed class Error {
         override val feilmelding: String,
     ):  Error(),
         MutationNyStatusSak.NyStatusSakResultat,
-        MutationSoftDeleteSak.SoftDeleteSakResultat,
-        MutationHardDeleteSak.HardDeleteSakResultat,
+        MutationDelete.SoftDeleteSakResultat,
+        MutationDelete.HardDeleteSakResultat,
         MutationKalenderavtale.NyKalenderavtaleResultat,
         QuerySak.HentSakResultat,
         MutationNesteStegSak.NesteStegSakResultat,
