+++ app/src/main/kotlin/no/nav/arbeidsgiver/notifikasjon/produsent/api/ProdusentAPI.kt
@@ -42,15 +42,12 @@ object ProdusentAPI {
 
                 QueryNotifikasjoner(produsentRepository).wire(this)
                 QuerySak(produsentRepository).wire(this)
-                MutationHardDeleteSak(hendelseDispatcher, produsentRepository).wire(this)
-                MutationHardDeleteNotifikasjon(hendelseDispatcher, produsentRepository).wire(this)
                 MutationNyBeskjed(hendelseDispatcher, produsentRepository).wire(this)
                 MutationNyOppgave(hendelseDispatcher, produsentRepository).wire(this)
                 MutationOppgaveUtfoert(hendelseDispatcher, produsentRepository).wire(this)
                 MutationOppgaveUtgaatt(hendelseDispatcher, produsentRepository).wire(this)
                 MutationOppgaveUtsettFrist(hendelseDispatcher, produsentRepository).wire(this)
-                MutationSoftDeleteSak(hendelseDispatcher, produsentRepository).wire(this)
-                MutationSoftDeleteNotifikasjon(hendelseDispatcher, produsentRepository).wire(this)
+                MutationDelete(hendelseDispatcher, produsentRepository).wire(this)
                 MutationNySak(hendelseDispatcher, produsentRepository).wire(this)
                 MutationNyStatusSak(hendelseDispatcher, produsentRepository).wire(this)
                 MutationKalenderavtale(hendelseDispatcher, produsentRepository).wire(this)
